backend/requirements.txt
Django==5.1.7
asgiref==3.8.1
django-environ==0.12.0
djangorestframework==3.15.2
pi==0.1.2
pip==23.2.1
psycopg==3.2.6
psycopg-binary==3.2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
langchain-core==0.3.47
drf-spectacular==0.28.0